requirements.txt
git+git://github.com/cbmi/avocado.git
django-preserialize>=1.0.6,<1.1.0
restlib2>=0.4.1,<0.5.0
python-memcached>=1.48
django-objectset>=0.2.3
